# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/donaciones/{id}": {
      "get": {
        "operationId": "DonacionesController_detalle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-13 · Un aporte: en que gastos se uso, con su evidencia",
        "tags": [
          "donaciones"
        ]
      }
    },
    "/api/v1/donaciones/{id}/constancia": {
      "get": {
        "operationId": "DonacionesController_constancia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-DE-07 · Constancia de una donacion confirmada, en PDF",
        "tags": [
          "donaciones"
        ]
      }
    },
...
    "/api/v1/suscripciones/{id}": {
      "patch": {
        "operationId": "DonacionesController_cambiarSuscripcion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-08 · Pausar, reanudar o cancelar en un clic",
        "tags": [
          "donaciones"
        ]
      }
    },
    "/api/v1/webhooks/pasarela": {
      "post": {
        "operationId": "DonacionesController_webhook",
        "parameters": [
          {
            "name": "x-pasarela-firma",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Confirmacion de cobro firmada por la pasarela",
        "tags": [
          "donaciones"
        ]
      }
    },
...
    "/api/v1/remanentes/{id}/destinos": {
      "get": {
        "operationId": "CausasController_destinos",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-CF-11 · Fondos que pueden recibir mi saldo",
        "tags": [
          "causas"
        ]
      }
    },
    "/api/v1/remanentes/{id}/eleccion": {
      "post": {
        "operationId": "CausasController_elegir",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "RF-CF-11 · Elegir devolucion o traslado de mi saldo",
        "tags": [
          "causas"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ayni API",
    "description": "Micro-mecenazgo dirigido con trazabilidad total. MVP v1 sin IA: la verificacion la produce el Motor de Reglas v0, que respeta el mismo contrato de datos que AIni.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}